update hosted wallet for kt
