+++ crates/tbdex/src/http_client/exchanges.rs
@@ -15,6 +15,7 @@ use crate::{
 use reqwest::Method;
 use serde::{Deserialize, Serialize};
 use web5::dids::bearer_did::BearerDid;
+use crate::http::exchanges::GetExchangesResponseBody;
 
 #[derive(Clone, Default, Deserialize, Serialize, Debug, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -110,7 +111,7 @@ pub fn get_exchange(
         Some(access_token),
     )?
     .ok_or(HttpClientError::ReqwestError(
-        "get exchanges response cannot be null".to_string(),
+        "get exchange response cannot be null".to_string(),
     ))?;
 
     let mut exchange = Exchange::default();
@@ -145,7 +146,27 @@ pub fn get_exchange(
     Ok(exchange)
 }
 
-pub fn get_exchanges(_pfi_did: &str, _requestor_did: &BearerDid) -> Result<Vec<String>> {
-    println!("TbdexHttpClient::get_exchanges() invoked");
-    Ok(vec![])
+// pub fn get_exchanges(_pfi_did: &str, _requestor_did: &BearerDid) -> Result<Vec<String>> {
+//     println!("TbdexHttpClient::get_exchanges() invoked");
+//     Ok(vec![])
+// }
+
+
+pub fn get_exchanges(pfi_did: &str, requestor_did: &BearerDid) -> Result<Vec<String>> {
+    let service_endpoint = get_service_endpoint(pfi_did)?;
+    let get_exchanges_endpoint = format!("{}/exchanges", service_endpoint);
+
+    let access_token = generate_access_token(pfi_did, requestor_did)?;
+
+    let response = send_request::<(), GetExchangesResponseBody>(
+        &get_exchanges_endpoint,
+        Method::GET,
+        None,
+        Some(access_token),
+    )?
+    .ok_or(HttpClientError::ReqwestError(
+        "get exchanges response cannot be null".to_string(),
+    ))?;
+
+    Ok(response.data)
 }
